Add ShouldDispatchAfterCommit to events + use NotificationDeliveryService
🐛 Fix: Events were dispatched before DB transactions committed,
which could cause listeners to see uncommitted data.

Changes:
- All event classes now implement ShouldDispatchAfterCommit
  - SalesInvoiceCancelled
  - StockLow, StockUpdated
  - TransferCompleted
- Console commands use NotificationDeliveryService for consistency
  - RemindOverdueInvoices
- All event listeners updated to use NotificationDeliveryService
  - HandleSalesInvoiceCancellation
  - SendPurchaseInvoiceCreatedNotification
  - SendSalesInvoiceCreatedNotification
  - HandlePaymentReceived
  - HandlePurchaseReturnProcessed
  - HandleSalesReturnProcessed
  - SendLowStockAlert
  - UpdateStockCache

This ensures:
  - DB transactions commit before events are dispatched
  - Listeners always see the correct data
  - Consistent notification delivery via NotificationDeliveryService
  - Better error handling with try-catch blocks

+++ app/Events/Stock/StockUpdated.php
@@ -1,4 +1,5 @@
 <?php
+
 namespace App\Events\Stock;
 
 use Illuminate\Broadcasting\InteractsWithSockets;
@@ -12,17 +13,22 @@ class StockUpdated implements ShouldBroadcast
     use Dispatchable, InteractsWithSockets, SerializesModels;
 
     public int $productId;
+
     public int $warehouseId;
-    public int $oldQuantity;
-    public int $newQuantity;
+
+    public float $oldQuantity;
+
+    public float $newQuantity;
+
     public string $operation;
+
     public string $updatedBy;
 
     public function __construct(
-        int $productId, 
-        int $warehouseId, 
-        int $oldQuantity, 
-        int $newQuantity,
+        int $productId,
+        int $warehouseId,
+        float $oldQuantity,
+        float $newQuantity,
         string $operation = 'update',
         ?string $updatedBy = null
     ) {
